Move markdown renderer
