+++ src/components/markdown-renderer/markdown-renderer.scss
@@ -1,4 +1,4 @@
-@import '../../../../node_modules/github-markdown-css/github-markdown.css';
+@import '../../../node_modules/github-markdown-css/github-markdown.css';
 
 .markdown-body {
   position: relative;
