+++ src/components/markdown-renderer/markdown-renderer.tsx
@@ -27,13 +27,13 @@ import ReactHtmlParser, { convertNodeToElement, Transform } from 'react-html-par
 import { Trans } from 'react-i18next'
 import MathJaxReact from 'react-mathjax'
 import { useSelector } from 'react-redux'
-import { TocAst } from '../../../external-types/markdown-it-toc-done-right/interface'
-import { ApplicationState } from '../../../redux'
-import { InternalLink } from '../../common/links/internal-link'
-import { ShowIf } from '../../common/show-if/show-if'
-import { ForkAwesomeIcons } from '../editor-window/tool-bar/emoji-picker/icon-names'
-import { slugify } from '../markdown-toc/markdown-toc'
-import { RawYAMLMetadata, YAMLMetaData } from '../yaml-metadata/yaml-metadata'
+import { TocAst } from '../../external-types/markdown-it-toc-done-right/interface'
+import { ApplicationState } from '../../redux'
+import { InternalLink } from '../common/links/internal-link'
+import { ShowIf } from '../common/show-if/show-if'
+import { ForkAwesomeIcons } from '../editor/editor-window/tool-bar/emoji-picker/icon-names'
+import { slugify } from '../editor/markdown-toc/markdown-toc'
+import { RawYAMLMetadata, YAMLMetaData } from '../editor/yaml-metadata/yaml-metadata'
 import { createRenderContainer, validAlertLevels } from './container-plugins/alert'
 import { highlightedCode } from './markdown-it-plugins/highlighted-code'
 import { linkifyExtra } from './markdown-it-plugins/linkify-extra'
