+++ src/components/markdown-renderer/regex-plugins/replace-legacy-gist-short-code.ts
@@ -1,4 +1,4 @@
-import { RegexOptions } from '../../../../external-types/markdown-it-regex/interface'
+import { RegexOptions } from '../../../external-types/markdown-it-regex/interface'
 
 const finalRegex = /^{%gist (\w+\/\w+) ?%}$/
