+++ src/components/markdown-renderer/regex-plugins/replace-legacy-slideshare-short-code.ts
@@ -1,4 +1,4 @@
-import { RegexOptions } from '../../../../external-types/markdown-it-regex/interface'
+import { RegexOptions } from '../../../external-types/markdown-it-regex/interface'
 
 const finalRegex = /^{%slideshare (\w+\/[\w-]+) ?%}$/
