+++ src/components/markdown-renderer/regex-plugins/replace-legacy-youtube-short-code.ts
@@ -1,4 +1,4 @@
-import { RegexOptions } from '../../../../external-types/markdown-it-regex/interface'
+import { RegexOptions } from '../../../external-types/markdown-it-regex/interface'
 
 export const replaceLegacyYoutubeShortCode: RegexOptions = {
   name: 'legacy-youtube-short-code',
