+++ src/components/markdown-renderer/replace-components/highlighted-fence/highlighted-code/highlighted-code.scss
@@ -1,5 +1,5 @@
 .markdown-body {
-  @import '../../../../../../../node_modules/highlight.js/styles/github-gist.css';
+  @import '../../../../../../node_modules/highlight.js/styles/github-gist.css';
 }
 
 .markdown-body pre code.hljs {
