+++ src/components/markdown-renderer/replace-components/highlighted-fence/highlighted-fence-replacer.tsx
