+++ src/components/markdown-renderer/replace-components/image/types.ts
