+++ src/components/markdown-renderer/replace-components/pdf/pdf-frame.scss
